refactor(tag): move category helper into tag domain
Relocate getPrimaryCategory and its test from lib/utils to the
tag domain so all tag-related helpers live in one place.
